requirements.txt
aiofiles==23.2.1
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
babel==2.16.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
docutils==0.21.2
executing==2.1.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fsspec==2024.12.0
furo==2024.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
polars==1.20.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-autodoc-typehints==3.0.1
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
